+++ src/main/java/com/github/cowwoc/tokenbucket/internal/ConsumptionFunction.java
@@ -1,7 +1,7 @@
 package com.github.cowwoc.tokenbucket.internal;
 
-import com.github.cowwoc.requirements.annotation.CheckReturnValue;
 import com.github.cowwoc.tokenbucket.ConsumptionResult;
+import com.github.cowwoc.tokenbucket.annotation.CheckReturnValue;
 
 import java.time.Instant;